LibGfx: Join SkiaFont.cpp into Font
No need to keep this function in a separate file.
